Add a confirmation step against accidental drops to /cs drop.
The procedure is that the user executes /cs drop <#channel>,
gets a key and then executes /cs drop <#channel> <key>.
Keys remain valid for 5-10 minutes and do not convey
any privilege. They are not stored in the channel
registration.

Currently, this also affects xmlrpc and is not
explicitly documented.

FDROP is not affected.