Make per-account vhosts a fallback for per-nick vhosts.
Due to this, there should not be a problem between
nickserv/vhost and hostserv/* anymore.

The GROUP and VHOSTALL commands set a per-account vhost
and clear all per-nick vhosts.

The request system has not been changed and only sets
per-nick vhosts.

Note that it is not possible to override a per-account
vhost with no vhost for a particular nick.