ircnet: Implement server hostmasking.
This is a feature of ircnet ircd where multiple servers appear
under the same name (for example *.de). Most commands important
for synchronization use the SID, which is different as usual.
An exception is KILL's prefix, server kills from masked servers
will therefore not work.

Also add the SID to the server_delete() log message (this is
in the network category). This is to disambiguate between
different masked servers but can also be useful more generally.