Add server_add and server_delete hooks.
Note that these are called for all affected servers,
not just the affected link.

Neither hook can prevent the action or otherwise destroy
the server objects. Due to differences in SQUIT semantics
across ircds I only plan to do this if there is a use case
for it.

There is no support for SQUIT reasons at this time
but server_delete takes a struct to allow this later.