Remove memo sending from /hs activate since there are many ways a user can see that their vhost has been approved and this can be quite spammy.
Keep it in /hs reject though so the user can see their vhost was
rejected if they're offline so they don't think that opers keep
forgetting to check the vhost queue.

Also, in my experience/opinion, vhosts are rejected a lot less often
than they are approved so this will still cut down on "snoop" spam
quite a bit.