Make ChanServ stay in "inhabit" channels with 1 other user indefinitely.
The "inhabit" flag is set if ChanServ joins to reset a
channel's TS or if it kicks the only user in reaction to
a join. It causes ChanServ to stay in empty channels
until a timer expires. Change this to also stay in
channels with exactly one other user, indefinitely,
preserving the inhabit flag until this changes.
This ensures autocycle scripts do not trigger and
the channel remains open for a while after kick on
split riding triggered.

ChanServ will part when a second user joins.

Make a similar change to BotServ bots.