Add the nickserv/set_enforcetime command/module. This allows users to set how long they'd like the delay before the enforcement occurs.
Some users may be slow to identify (if the default delay is short) so there's
that and some users might use CertFP or SASL or a autoidentify script so they
can set it to something like 1-5 seconds so people can't hang on their nicks
as long. This is the cleanest way to do it and account for all possibilities/types
of users.