Many fixes to chanserv/access and hide_xop.
Including:
Automatically add CA_FLAGS if adding CA_FOUNDER to a role or access.
Checked if a role exists prior to assigning it to a CA's level.
Removed ability for people to have multiple roles or templates wth the same flags.
If the last founder's access is removed, it will search for a successor.
Enabling hide_xop now truly removes the default xOP templates.