Fix auto-unbinding of commands following 482d4cdf7f75
The previous implementation stored references to the service and command
objects, for later unbinding. The aforementioned commit stopped this from
working. To fix, change Atheme::Command to be owned by the Perl code, so that
it can be stored persistently, and store the service name instead of an object
reference.