libathemecore: add object_sink_ref()
This is for dynamically allocated objects that should start their lifecycle
with refcount = 0.  An example of such an object would be a dynamically
allocated entity tied to one or more channel access entries.  When the entity
is removed from all channels it has access to, the entity needs to be disposed
by the object system.  Starting the object's lifecycle with a refcount of zero
allows this to happen.