sasl: Add new-style EXTERNAL authentication.
This uses an additional parameter in the S message to send the
certfp. The certfp is sent as 40 lowercase hexadecimal digits
without base64.

To comply with the SASL RFC for the EXTERNAL mechanism, a
client response is required, which must be empty or match
the name of the target account.