Fix NLS on FreeBSD and other systems with separate libintl.
Something (probably GNUish) moved the libraries for gettext to a separate
variable which we do not use. This breakage cannot be noticed on Linux,
where gettext is part of glibc.

Note that NLS on FreeBSD still requires passing CFLAGS=-I/usr/local/include
LDFLAGS=-L/usr/local/lib to configure.