libathemecore: make chanacs_change() work with literal targets.
chanacs_find() is a fuzzy matcher, which will return a chanacs_t entry for an entity
if it is matched by any conditional, such as group or exttarget qualification.  we should
instead use chanacs_find_literal() because we want to make an edit to the literal channel
ACL when calling chanacs_change().

Closes: SRV-117