botserv: Make sure botserv bots join and inhabit before kicking a user on join.
Put bs_join before cs_join and treat botserv channels as guarded channels
for the purposes of the restricted, akick and mlock +i checks.

If changets was not in use, this could cause crashes or chanserv joining
instead of botserv.

With this fix, it is no longer needed to refcount channel_t objects; also,
this fix prevents kick-rejoin loops.