protocol/plexus: *plonk*
nobody cares about rizon's ircd and it is kind of like ircd-seven in that really
nobody should use it.  it's not like rizon is going to ever be running atheme, and
they contributed the module to begin with.

plus, the only network running it with atheme is adminned by e-trash so that's always
cool.