libathemecore: call shutdown(2) on connections being closed
This allows us to hint to the kernel that we need to cleanly tear down the
associated TCP session for the socket.  On some systems (*ahem* Linux),
without this hint we get a stale leftover connection on the IRCd side which
occasionally causes some level of chaos on services restart.