Combine all chanacs entries matching a user to determine their access.
This means that having an entry for a specific user no longer overrides
entries for a group that user is in, but the combined flags apply.

The ChanServ WHY command already pretended this was the case.

Note that chanacs_find() is rarely useful. In most cases either
chanacs_find_literal() or chanacs_entity_has_flag() are needed. One
exception is finding an akick reason to kick with, since this needs
any matching chanacs entry with +b, it does not matter much which.