Add a new hook channel_succession to remove +S from the new founder properly.
The channel_succession hook is called when it is certain which user
will become the new founder and is useful to commit any changes resulting
from the succession, apart from setting the new founder as founder which
is done by the core.

The channel_pick_successor hook should not commit any changes resulting
from the succession; a subsequent hook function might override the
current function's decision.