Actually send the new vhost
This bug was introduced in 35cf1b, which assumed that u->vhost had
already been updated with a stringref-ed version of the new host.  As
user_sethost() in libathemecore does this for us, just pass through the
new host to it, after checking that it's different from the current
vhost.