Change the +u (unban) to the +e (exempt) flag.
The +r flag retains all the commands it had, including unban, and access
to add/remove akicks with +f, but loses the akick exemption.

The +e flag allows unban/unquiet on self only (so only via IRC) and grants
exemption to akicks.

There is code to convert old +r to new +re but not code to convert old
+u to new +e.