cmode: Map mode +l 0 to mode +l 1.
Some ircds like inspircd allow a limit of 0 which allows nobody in.
This was treated as no limit by some parts of services, which is
clearly undesirable (for example, mlock -l did not revert a +l 0
change).

I think +l 0 should not be allowed so this commit changes it to +l 1
which has almost the same effect.