protocol: Fail extban check if mask contains space
In certain circumstances, the mask string passed into is_extban can
contain space(s) and currently there is no validation in any protocol
module implementing extbans that the input mask indeed contains no
spaces. An example circumstance is ChanServ's BAN command. The following
command will cause ChanServ to set a ban on the first mask due to how
the MODE command works, but fail to add the ban to Atheme itself due to
the mask being the full rest of the string as opposed to just a single
word.

`/msg ChanServ BAN #channel $a:mask1 $a:mask2`

This patch adds a simple strchr check for the space character to each
extban check, fixing the issue outlined above.