[database] Make services respect an external umask when saving
Currently, services sets a hard umask of 0077 when it starts up,
and uses fopen(3) to write out the database file.

This commit removes the umask() call and makes the database file
0660 by default; should you desire less permissions on it (such
as 0600), you can start atheme with an appropriate umask (0077),
or simply not give read/write access to the directory that the
database file is written to, etc.