protocol/inspircd: Only set hideoper mode on oper pseudoclients
This fixes an issue when Atheme is used with an InspIRCd network with
the hideoper module loaded. InspIRCd adjusts the number of online
opers shown in reply to LUSERS according to the number of hidden opers
(+H clients). This breaks if non-oper clients have the (otherwise
oper-only) hideoper mode set. (Plus it also does not make much sense
to set the hideoper mode on non-opers.)

This patch changes the protocol module so that the hideoper mode is
only set on pseudoclients which are also opered.

Closes #478