email templates: Fix leading whitespace
The leading whitespace is barely visible in many email clients
and was often cause for sending the message directly to the channel
for clients that do not interpret " /command" as a command. This
was most notable for qwebirc.

(cherry picked from commit 72821fa6eea3e7a20c0e3286f709fe6cc0e214f5)