nickserv/register: Remove password from registration confirmation
The password should not be sent back to the client for security reasons.

An attacker could potentially read the message from the screen until the
message has been cleared (e.g. when registering in public) or the
message could be logged by the client.

In the latter case, this might not even be known to the user.

As for the former case, many applications obscure the password to
prevent such situations from happening. Since the password is not really
necessary in this case, we can also leave it out.

Typos that go unnoticed during registration can be worked around by
using SENDPASS for now, in the future it might be worth to add a special
VERIFY case to have the user enter the password twice.

Closes atheme/atheme#577 ("When registering the response contains the
password as plain text")