myuser_login: remind users of email verification
This removes the similar check from nickserv/identify so other methods
of identification will also remind users (notably, SASL).

The chosen message also spans three lines, similar to the message given
when registering an account, plus includes a bold uppercase part
(similar to nickserv/info output), in an attempt to draw the user's
attention to the fact that they are using an account that will soon
expire if they fail to act on the verification email.

The notion of sending this message from a hook in nickserv/verify was
briefly discussed in #atheme, but it seems to fit myuser_login rather
well. Additionally, this avoids hardcoding NickServ as source of the
message when SaslServ might send the other on-login messages while not
requiring a rework of the on_user_identify hook parameters.

Closes #566.