modules/crypto/pbkdf2v2: add support for HMAC-SHA1
The master branch (what will become 7.3) got this too, but it was
added there for future possibility of implementing SASL SCRAM support.

We will not be implementing support for SCRAM in version 7.2, but
people who used the pbkdf2v2 module with SHA1 in version 7.3 might
want to downgrade to version 7.2, so we need support for verifying
those hashes too.