Revert "libathemecore: add constant-time memory comparison function"
This function compiles well and behaves correctly under all of my
tests, but on the advice of another security-focused software
developer it seems some whacky compiler out there is going to be able
to break it under some circumstances, or even worse, optimise it out.

It was added for password crypto modules, but the avalanche effect
makes comparison timing info-leaks a non-issue in the context of
password hashing anyway... these aren't MACs.

This reverts commit 1da9c8ce925bae6fcac90a150ed97d8ea74e67e8.