modules/crypto/pbkdf2v2: fix decoding bug for SHA2-512
siniStar found a bug in our base64 decoder, in that it will fail
when asked to decode 64 bytes of output into a 64-byte buffer,
but succeed if given a 65-byte buffer.

One would expect this behaviour from an encoder, since the result
of encoding is a printable string which needs to be NULL-
terminated (thus requiring an extra byte), but it should not
happen for decoding.

Until the bug in the decoder is identified and resolved, this
will work around the issue.

TODO: Revert this commit