modules/saslserv/scram-sha: adjust buffer size for base64 encoding
When converting a user's regular HMAC PBKDF2 password hash into a
SCRAM one, make sure our buffer for encoding ServerKey and StoredKey
is big enough. This is not currently an issue because SCRAM-SHA-512
has not been standardised and this module does not support it, but
that may change in future.