modules: fix up declarations
While investigating something unrelated the other day, I stumbled
upon this quirk in module declarations.

The second field of DECLARE_MODULE_V1() was supposed to mean
"is a core module?" (can't be reloaded), or so I thought.

It turns out that this field is used to initialise the 6th field
of the MAPI module header symbol, and that field is an
'module_unload_capability_t' enumeration, not a boolean.

The compiler doesn't warn for this because bool in C is a kludge,
so when you're passing 'false' you're really passing '0', which
matches MODULE_UNLOAD_CAPABILITY_OK, and when you're passing
'true' you're really passing '1', which matches
MODULE_UNLOAD_CAPABILITY_NEVER.

Still, it is a discontinuity that should be addressed before it
confuses someone else for a moment too.