modules/crypto/pbkdf2v2: implement SASLprep password normalization
This is only used in SCRAM mode (for SASL SCRAM-SHA).

Furthermore, warn the administrator if verification against a SCRAM format
hash is being performed but GNU libidn is unavailable.

Furthermore, disallow the administrator from configuring the module to
compute new password hashes in SCRAM mode if GNU libidn is unavailable.