modules/crypto/pbkdf2v2: allow and enforce salts to be base64-encoded
The last remaining concern with this module was that it used salts
composed of only ASCII characters. Off the top of my head, this has 2
disadvantages:

1) We cannot verify PBKDF2 hashes imported from another program
   (e.g. during a services database migration from competing software)
   if that program does not also use ASCII salts

2) The available entropy in our salts was less than their length
   would indicate to a casual observer; there were only 2^96-32
   possible combinations instead of 2^128, i.e. approximately 96 bits of
   entropy instead of 128 bits, for a "128-bit" salt

Item #2 is more a technical matter as salts only need to be unique,
not random or unpredictable, and 96 bits is already far into this
territory. Item #1 however may be a source of concern in the future.

This commit processes base64-encoded salts, generates base64-encoded
salts, and forces base64-encoded salts (during verification of an
existing older password hash, it will be rehashed if successful).

Compatibility is retained for the older format, to make upgrades
painless.

This functionality will also be backported to the release/7.2 branch, in
case people start using this new functionality now and then decide to
downgrade to that version in the future.