modules/saslserv/main: sasl_packet: use larger buffer
The new SASL SCRAM-SHA can generate large responses (particularly when
the client supplies a large nonce) which could be base64-encoded to a
length exceeding the current value.

Use the same maximum length for outgoing data as the restriction we
place on incoming data (and remove that magic number too).