modules/saslserv/main: fix (dead) potential segfault in moddeinit
All of the SASL mechanism modules import a symbol from the
saslserv/main module when they are loaded, which puts
saslserv/main into their dependency tree.

If saslserv/main is unloaded, all of the modules that depend on
it are unloaded first, and those modules are responsible for
unregistering the mechanism(s) they added. This will result in
any sessions using that mechanism also being destroyed.

If there are still sessions left when saslserv/main is being
unloaded, it must be because one of the mechanism modules did
not unregister itself.

However, the moddeinit function then went on to try and clean up
the sessions itself. The problem is that destroy_session() will
call ->mech_finish() (if it is provided by the module) but the
module that contains that function is now unloaded, so we will
be calling a function that no longer exists.

This isn't actually possible if people don't write their own
SASL mechanism modules, because all of the in-tree modules
correctly unregister themselves when they are being unloaded.

Remove the loop and upgrade the existing warning to a higher log
level so that it is seen by more users.