modules/saslserv/main: avoid TOCTOU and make user_can_login hooks useful
If a session that requires more than one step (ECDSA-NIST256P-CHALLENGE,
SCRAM-SHA) is initiated, the username and authzid will be stored into
the session structure as they become available, but the account(s) could
be dropped by the time the session authentication succeeds.

Normally this would fail login, unless an account with the same name was
recreated. At that point, the user would be logged into the new account,
even though they did not have its credentials.

The impact of this bug is considered very minor (sessions time out after
60 seconds if they don't make definitive progress, and the unwary user
would have to register an account with a name of a just-dropped account
within that time window).

Fix this by storing the entity IDs into the session structure instead.
Accounts always get new entity IDs as they are created, and they are
not reused.

Also in this commit is the moving of user_can_login() hook calls from
login_user() to immediately when the authcid/authzid become known by the
SASL mechanism module(s).

This is motivated by the fact that the new SCRAM-SHA module allows pass-
word logins (in a sense), and a module installed that hooks this to deny
password logins via SASL in an effort to prevent bruteforcing them would
be ineffective: the SCRAM-SHA module sends the server-final-message if
the password was correct, and not otherwise. Thus, the presence of a
server-final-message value in an SASL SCRAM-SHA exchange confirms that
the password was correct, even though the exchange is later aborted by
the hook. Fix this by calling the hook immediately.