modules/saslserv/*.c: adjust function parameter types
The input data to ->mech_step() functions is base64-decoded data;
it could be arbitrary and the API should not assume it is a string.
Additionally, the function should not modify the memory it points to.

Similarly, the output data could be arbitrary bytes (and in the case
of e.g. ecdsa-nist256p-challenge, is), not a string, and it's going to
be base64-encoded anyway.

For these reasons; use void, not char, and make input const.