modules/saslserv/main: simplify SASL 'S' command processing
The first AUTHENTICATE command's argument is the mechanism name,
so we don't need to worry about buffering data from the client
in this case because the maximum length of a mechanism name is
less than the maximum length of data a client can send at once.

This avoids unnecessary memory allocation.