modules/saslserv/*.c: add new kind of start/step mech error code
ASASL_FAIL results in a bad_password() on the user, even if the
failure was due to some kind of internal error (e.g. too small a
buffer passed to base64_encode()), which is undesirable.

Add a new code (ASASL_ERROR) that aborts the sessions all the
same but does not bad_password() the user, and reserve ASASL_FAIL
only for cases where the client supplied correctly-formatted data
but nonetheless invalid credentials.