Build System: Fix up OpenSSL mess
Upstream libmowgli uses --with(out)-openssl, so let's use that here too.

This means we no longer need to execute the rather horrendous
`./configure --disable-ssl --without-openssl' to truly build without it.

Secondly, while libmowgli uses OpenSSL proper, Atheme only uses the
libcrypto library. We have no need to link against -lssl, so let's not.

Thirdly, if we're going to try building against OpenSSL, test that the
3 universally-required header files (evp.h, hmac.h, sha.h) are available,
and remove unnecessary header includes from the code using OpenSSL.

Finally, if the build system detects that OpenSSL is unavailable but
GNU libidn is, disable the latter because GNU libidn is only used for
the SASL SCRAM-SHA functionality which needs OpenSSL too. If the user
explicitly gave --with-libidn, warn them about it.