Password Crypto API: Several significant improvements
There were 4 concerns with the codebase in its current state:

When a password is verified against a password hash by a password
crypto module, and the verification fails, libathemecore would go on
to try the next module(s) even if the password hash was not produced
by those modules. This is inefficient.

When a password crypto module provides a 'verify' function,
libathemecore/auth.c and libathemecore/crypto.c will only use the
'crypt' function to encrypt new passwords or re-encrypt existing ones,
but it will first call the 'salt' function to generate a salt string,
only to immediately give it back to the 'crypt' function so that it
can parse the parameters from that string. This is inefficient.

When a password is verified against a password hash by a password
crypto module, and that module provides a 'recrypt' function to
indicate to libathemecore/auth.c that the password should be re-
encrypted, the 'recrypt' function has to be invoked and parse the
parameters again (which 'crypt' or 'verify' already did). This is
inefficient.

Existing callers of crypt_string() now pass NULL as the salt argument,
rendering that argument completely useless. In addition, crypt_string
only tried the first encryption-capable crypto module, and if that
failed, gave up entirely. This is undesirable. It is better to fall
back to the next module than to not encrypt a password.

These defects are resolved as follows:

The 'salt' funcptr has been removed from the crypto modules. Those
modules are now expected to generate their own parameters on-the-fly
if their 'crypt' function is passed a NULL argument as the salt. This
avoids encoding parameters into a string only to immediately scan the
parameters from that string.

The 'recrypt' funcptr has been removed from the crypto modules. Those
modules now indicate whether the password is to be re-encrypted by
setting a flag in their 'verify' function. This avoids parsing twice.

Modules that provide a 'verify' function can indicate to libathemecore
that the password hash was indeed produced by this module, so that if
verification fails, libathemecore does not continue trying.

A new function crypt_password() has been created that takes only one
parameter; crypt_string() has been retained for compatibility and is
simply a wrapper around crypt_password(). crypt_string() now ignores
its second argument. crypt_password() will try each encryption-capable
crypto module in sequence, instead of only the first such module, in
the hope that at least one of them will be able to encrypt it.