PBKDF2v2/SCRAM-SHA: Fix configuration processing bug
When the SCRAM-SHA module is loaded, it registers a hook for the
config_ready event, to decide (based on the value of pbkdf2v2::digest)
which SASL mechanism to register.

This works fine during startup; however if modules/saslserv/scram-sha is
reloaded, the hook no longer fires and the module registers no
mechanisms until services is rehashed.

Fix this as follows:

- modules/crypto/pbkdf2v2 adds the hook for config_ready instead
- modules/crypto/pbkdf2v2 sets the config variables to zero by default
- When the hook is executed, the variables will be set to their default
  values if they have not already been configured (this would be during
  startup, or if the configuration options are missing from the file)
- modules/saslserv/scram-sha gives a function pointer to the pbkdf2v2
  module, to be called when that module is (re)configured

Initially I thought the only way to fix this would be to move the
pbkdf2v2{} configuration processing code to libathemecore/conf.c, but
this is a better approach.

Reported-by: Mantas Mikulėnas <grawity@gmail.com>