libathemecore/explicit_bzero.c: prevent compiler removing function call
When compiling this program with a sufficiently advanced compiler in
full LTO mode, and we don't have explicit_bzero(3) or memset_s(3), the
compiler is free to remove the call to explicit_bzero(3) as part of
Dead Store Elimination, defeating the point of using it.

The issue is that under these conditions, explicit_bzero.c boils down to
1 static volatile function pointer and 1 function which only calls it.

The compiler is able to prove that the value of the function pointer will
never change, even though it is volatile, because it is isolated to this
compilation unit, and nothing in this compilation unit changes its value.

To solve this problem, we make it extern, with a forward declaration as
such. Because this file is compiled and linked into a library, we can now
theoretically have our value adjusted by a consumer of this library,
which the compiler can't prove doesn't happen.

This results in Clang 6.0 with Thin LTO no longer removing the calls to
explicit_bzero.c:explicit_bzero() as part of Dead Store Elimination.

While we're here, if we don't have explicit_bzero(3) but we do have
memset_s(3), define a macro equivalent. This avoids the overhead of a
function call for something that could be efficiently inlined.