modules/saslserv/scram-sha: provide bounds for client nonce
At the moment, a client can send a nonce as large as their
permitted maximum client-to-server SASL payload (8192 bytes),
minus at least 6 bytes, and the server will have no room to
append its nonce.

This will result in an error message in the logs, and is a
cheap way for malicious clients to flood the log.

Correct this by putting a smaller upper bound on their nonce
length.