modules/crypto/pbkdf2v2: fix computation of new digests
When we're encrypting a new password, we generate a random salt for it,
but we neglected to set the salt length before running the computation.

This results in encrypting the password with no salt.

When it comes time to verify it, the salt is decoded and its length is
set, which results in computation of a different digest.

This regression was introduced in commit 317796ead0ea5f34d4a9.