libathemecore: implement digest algorithm agnostic interface
This commit converts the existing MD5 implementation.

This supports the following operations:

- Digesting data (all in one shot)
- Digesting data (initialization, adding data, and final computation)
- HMAC (in both of the above modes)
- PBKDF2-HMAC

It also adds an internal testsuite to verify the correctness of
the above modes of operation with MD5, and finally converts the three
existing callers of the old MD5 implementation to use the new one:

- modules/crypto/ircservices.c:atheme_ircservices_encrypt()
- modules/crypto/rawmd5.c:atheme_rawmd5_verify()
- libathemecore/function.c:create_challenge()

Finally, it removes the old MD5 implementation.