libathemecore: add Nettle support to digest API
This is currently sub-optimal; I could not get HMAC mode of operation
to work using libnettle's own HMAC functionality, so I have copied the
internal frontend and adjusted it for libnettle appropriately instead.

As an aside, it uses less memory anyway, so it may be worth just keeping
it this way for the forseeable future.