Travis CI: Attempt some minor improvements
- Use Clang 5.0 instead of 6.0 because the LLVM developers bumped the
  unstable branch to package version 7 which isn't whitelisted and the
  new branch that contains version 6 isn't whitelisted either.

- Don't waste time fetching clang packages for gcc builds.

- Verify that 'make install' works too.

This does introduce some configuration duplication (packages, env) but
may be worth it for the time savings.

I don't know if this is going to work or not; I tried the "Trigger
Build" function in the administrative interface but it was all mangled
up with the current (at the time) configuration as well:

  <https://travis-ci.org/atheme/atheme/builds/331388039>

If this doesn't work I'll revert this commit.