Makefiles: A few more cleanups
Mainly we separate LDFLAGS from LIBS some more.

For the Perl API, we don't unconditionally assume we are building
against the in-tree libmowgli. This may break if we build against
a system libmowgli installed into a non-standard location because
we may not find its installed headers, and will use the ones from
the submodule instead.