libathemecore/arc4random.c: a few more improvements
- Use ARM mbedTLS as the RNG directly if it is available
- Rip out the OpenSSL code that obtain seeds for the ChaCha20 RNG

This simplifies things somewhat: if we have a high-quality RNG
library, we just use that, otherwise we stick with the ChaCha20 RNG
from OpenBSD and the seed is gathered only from the operating
system (via getentropy(2), getrandom(2), or urandom(4)).

This results in arc4random() and arc4random_uniform() being generic,
both calling arc4random_buf() which have different implementations
based upon whether we are using the built-in ChaCha20 code or not.

Note that this file is still only compiled and used if the user
does not have arc4random*(3) already.