groupserv/main: write out group definitions before their access lists
If a group '!a' includes an access list entry for group '!b' (which is
lexicographically after) then upon read-back (when restarting services)
the access list contains an entry for a (for the moment) non-existent
group which is ignored. This results in data loss.

The fix is simple: Write out all group definitions (& their flags and
metadata) and then all group access lists.

Reported-by: Samuel Hoffman <miniCruzer@live.com> (sjh)
Reported-by: Aaron M. D. Jones <aaronmdjones@gmail.com> (amdj)
Reviewed-by: Janik Kleinhoff <janik@kleinhoff.de> (ilbelkyr)