libathemecore: generate saner challenges
Now that we have a message digest API that supports HMAC, just
use our 2 input parameters (source name, thing-being-dropped
name) as the key and data.

This avoids generating challenges with a time component and all
of the extra complexity of validating just-expired challenges.

It also makes clear that this is not supposed to be used for
security purposes; the function was renamed to contain 'weak'.

It also now detects errors and returns a failure appropriately.