modules/crypto/posix: replace with better crypt(3) modules
The algorithms that crypt(3) implements are platform-specific, and
only one of them (DES) can be considered "standard". It may fail,
or it may return a different result to that of other platforms.

For this reason, these new modules include a self-test routine,
which will verify that the algorithm works before allowing itself
to be used.

They are also now verify-only, because both DES and MD5 are very
weak algorithms.

Closes #603