modules/groupserv/main/groupserv_common.h: s/(mygroup)_t/struct \1/
Types with names ending in '_t' are reserved and should not be defined
by applications.

The typedef is kept for the hook system and Perl scripts, however.
We should investigate renaming it at a later date.

I also moved mygroup and groupacs to include/account.h and privs
to include/privs.h. This made things simpler.