modules/: clear mod_deinit() if MODULE_UNLOAD_CAPABILITY_NEVER
Even if mod_init() fails (m->mflags (|)= MODTYPE_FAIL), mod_deinit()
is never called, because it isn't added to the `modules' list until
m->modinit() has run without errors, and module_unload() only invokes
m->deinit() if it is in the `modules' list.

This makes all code in mod_deinit() pointless if the module doesn't
perform any global modifications before failing; and it would never
get the chance to undo those changes anyway.

So, having a mod_deinit() at all in a module that cannot be unloaded
is pointless, regardless of whether mod_init() can fail or not, but
our simple declaration macros expect one, so leave it empty rather
than removing it.