include/Makefile: improve serno.h generation
This introduces a small shell script which only writes to serno.h if
it is necessary (i.e. if the commit hash stored within it at the moment
is not the same as the one we would write).

It also stores a shorter length hash (that is nonetheless useful for
identifying the specific commit and can be passed to git-show(1) and
such as normal).

This also means that every `make' now no longer recompiles
libathemecore/version.c and modules/chanserv/version.c, and no longer
relinks atheme-services. This should improve the reproducibility of
builds too (--enable-reproducible-builds).