include/crypto.h: make `struct crypt_impl' const
This should help prevent accidental or deliberate manipulation of e.g. the
verify funcptr which results in all passwords successfully verifying.

The reason is that making the structure a static constant in each crypto
module should convince the compiler to put it into a read-only section in
the module file. Writing to it would thus trigger a memory access violation.