include/: reindent all structures and other misc fixups
Members are now indented with a single tab, and then the data
type (which may be a pointer type), followed with spaces to the
next free column that is a multiple of 8 characters (assuming a
tab is 8 characters wide) before the member name.

To elaborate:

struct {
	const char *  (*foo)(void *);
	const char *    bar;
};

... is used, instead of:

struct {
	const char *    (*foo)(void *);
	const char *    bar;
};

i.e. the first member name starts at "f", not "(".

This makes determining the names of adjacent members and visually
enumerating the members of a structure much easier.