modules/saslserv/main.c: abort session if we receive '*' as data
We already fail to base64_decode() it and then abort the session anyway,
but this just saves time and avoids trying to decode it in the first
place, which will always fail and produce an error message in the debug
log.

cf. https://github.com/ircv3/ircv3-specifications/pull/298#issuecomment-271336287

Reported-By: James Wheare