modules/crypto/ircservices: apply commit 507f49117bea9a3f
This one is not a straightforward modification. While retagging this
module I discovered that the header on the original code in Anope and
the header in this module refers to its original COPYING file (in
Anope), which is the GNU General Public License (version 2). The
header also states that it is based on Epona, which is also GPLv2.

This makes the subsequent addition of our own header, referring to
our own doc/LICENSE file (which is the ISC license), illegal, as the
GPLv2 license is not compatible with the ISC license.

We cannot simply relicense GPLv2 code as ISC code and call it a day,
and any derivative modifications (such as adaptation to Atheme's
Password Crypto API) would also be licensed under the GPLv2.

There is a clear path forward, however. The Atheme portions of this
code were completely rewritten by me in 2017 and 2018 as part of my
substantial modifications to the Password Crypto API, and the Digest
API, respectively. This allows me to simply relicense the entire file
as GPLv2 only.

For the sake of completeness I have preserved the original ISC
copyright attribution to the previous Atheme Project developers (and
extended the year to the last substantial modification performed by
that group), even though it is not strictly necessary.

There is one caveat, however. Should a user load this module into
their Atheme installation at runtime, such an installation may
become GPLv2-licensed (this is up to lawyers to decide, and I do not
wish to play one).

To alleviate this concern, warranted or not, I have already modified
the master branch to put all of the legacy compat crypto modules
behind a configure flag, so they are not built or installed by default.