libathemecore: add dedicated memory zeroing function
This removes the old explicit_bzero(3) wrapper if it isn't available;
providing a new smemzero() function which chooses between memset_s(3),
explicit_bzero(3), or memset(3) through a volatile function pointer.

It also clarifies the situation regarding that pointer (see the
comment in smemzero()) and moves the pointer to another compilation
unit to make compilers more unlikely to remove any calls to smemzero()
or to remove the indirect call to memset(3) from smemzero().

Finally, it defines a macro at configure-time that is required to
properly detect memset_s(3) in the first place.