libathemecore: add libsodium support for memory allocation
This will replace our use of libc malloc(3) family of functions with
those provided by libsodium. It is disabled by default.

This is not recommended for production usage; it adds significant
virtual memory overhead and a small runtime cost for every allocation.
It is a compromise between not being able to reliably detect double-
frees or reallocs/frees of something not on the heap, and where the
problems occurred, and running the entire program under Valgrind
memcheck instead. It will also detect under- and over- reads and writes
of the allocations.

Only Atheme developers and experienced users suspecting memory
corruption bugs, or paranoid users, should enable this feature.